pyproject.toml
[project]
name = "websocket-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.6",
    "mypy>=1.14.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "securely>=0.1.9",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.34.0",
]
